fix test for v1.2 sample minimal las file
The minimal v1.2 doesn't have header names in the data section.
So the solution is to remove the hard coded 'headers = TRUE'
from data.table::fread. By default a header line will be created
if the first data line has all fields of type 'character'. This
solution also works with the v1.2 sample.las which has a header
data row.