Replace magic-quotes with regular double-quotes n the install command.
The `devtools::install\_github(“Gitmaxwell/lastools”)` instruction for lastools install contained magic-quotes that cause an error when cut-and-pasting the command to an `R>` terminal.  Replacing the magic-quotes with regular quotes enables the cut-and-paste to install lastools.  The related README.Rmd already has the correct double-quotes.

Let me know if this change could be accepted (or rejected) or
needs some additional changes before being approved and merged.

Thank you,
DC